backend/requirements.txt
Django>=5.0
psycopg2-binary
python-dotenv
djangorestframework
drf-nested-routers